Add GetSFIDsByExternalId which always returns a slice of ids
