+++ force/sobjects.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"net/url"
 	"strings"
+
+	"github.com/opendoor-labs/go-force/sobjects"
 )
 
 // Interface all standard and custom objects must implement. Needed for uri generation.
@@ -110,6 +112,41 @@ func (forceApi *ForceApi) DeleteSObject(id string, in SObject) (err error) {
 	err = forceApi.Delete(uri, nil)
 
 	return
+}
+
+func idsFromURIs(uris []string) []string {
+	ids := make([]string, len(uris))
+	for i, uri := range uris {
+		parts := strings.Split(uri, "/")
+		ids[i] = parts[len(parts)-1]
+	}
+	return ids
+}
+
+func (forceApi *ForceApi) GetSFIDsByExternalId(apiName, externalKey, externalId string) ([]string, int, error) {
+	uri := fmt.Sprintf("%v/%v/%v", forceApi.apiSObjects[apiName].URLs[sObjectKey], externalKey, externalId)
+
+	params := url.Values{"fields": []string{"Id"}}
+	sobj := sobjects.BaseSObject{}
+
+	statusCode, err := forceApi.Get(uri, params, &sobj)
+	if err == nil {
+		return []string{sobj.Id}, statusCode, nil
+	}
+
+	if statusCode != 300 {
+		return []string{}, statusCode, err
+	}
+
+	// We don't have access to the json that was returned, so make the
+	// same call passing in a slice to unmarshal into.
+	uris := []string{}
+	statusCode, err = forceApi.Get(uri, params, &uris)
+	if err != nil {
+		return []string{}, statusCode, err
+	}
+
+	return idsFromURIs(uris), statusCode, nil
 }
 
 func (forceApi *ForceApi) GetSObjectByExternalId(externalKey, externalId string, fields []string, out SObject) (statusCode int, err error) {
